requirements.txt
openai
backoff
python-dotenv
transformers
tensorflow